+++ web/src/api/core/OpenAPI.ts
@@ -21,7 +21,7 @@ export type OpenAPIConfig = {
 
 export const OpenAPI: OpenAPIConfig = {
   BASE: "http://localhost:8080",
-  VERSION: "0.4.1",
+  VERSION: "0.4.2",
   WITH_CREDENTIALS: false,
   CREDENTIALS: "include",
   TOKEN: undefined,